pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["leo-guinan <leo@copy.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.104.1"}
uvicorn = "^0.23.2"
langchain = "^0.0.330"
python-dotenv = "^1.0.0"
openai = "^0.28.1"
python-decouple = "^3.8"
pymongo = "^4.6.0"
chromadb = "^0.4.15"
sentence-transformers = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
